requirements.txt
flask 
Flask-Bootstrap4
flask-sqlalchemy
Flask-WTF
Flask-Bcrypt
email-validator
requests
flask_cors
pycryptodome
Flask-reCaptcha